requirements.txt
xmpppy
mqtt